+++ CMakeLists.txt
@@ -237,6 +237,7 @@ add_executable(${PROJECT_NAME}_person # detectPersonHOG
 )
 add_executable(${PROJECT_NAME}_darknet_debug # darknet_ros
   src/darknetImg/main.cpp
+  src/darknetImg/constructor.cpp
   src/darknetImg/methods.cpp
   src/darknetImg/debug_methods.cpp
 )
